// src/common/enums/offer-status.enum.ts
export enum OfferStatus {
  NEW = 'new',
  IN_PROGRESS = 'in_progress',
  PRICE_REVIEW = 'price_review',
  PENDING_CONFIRMATION = 'pending_confirmation',
  CONTRACT_READY = 'contract_ready',
  CONTRACT_SIGNED = 'contract_signed',
  PAYMENT_IN_PROGRESS = 'payment_in_progress',
  ITEMS_IN_PROGRESS = 'items_in_progress',
  IN_DELIVERY = 'in_delivery',
  DELIVERED = 'delivered',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled',
}